bridge_macros

Function are_args_valid

pub(crate) fn are_args_valid(
    original_item_fn: &ItemFn,
    params: &[Param],
    takes_env: bool,
) -> Result<(), Error>

Optional and VarArgs types are supported to create the idea of items that might be provided or for providing a list of zero or more items that can be passed in. The nature of optional and varargs are context dependent because variable numbers of arguments have to be at the end of the function signature. This method verifies that items marked as Optional are last, and VarArgs is supported but only in the last position, which can be after any number of Optional arguments. This means non Optional/VarArgs types must come before all Optional and VarArgs types.