bridge_macros

Function generate_intern_fn

pub(crate) fn generate_intern_fn(
    original_fn_name_str: &str,
    fn_name_ident: &Ident,
    fn_name: &str,
    doc_comments: String,
) -> TokenStream

write the intern_ function code. This code is generated to be called within sl-sh to avoid writing boilerplate code to submit a function symbol and the associated code to the runtime. Every builtin function must be inserted into a hashmap where the key is the name of the function and the value is a function expression that stores the name of the rust function to call and its documentation. It looks like the following in all cases: