bridge_macros

Function generate_sl_sh_fn

pub(crate) fn generate_sl_sh_fn(
    original_item_fn: &ItemFn,
    attr_args: AttributeArgs,
) -> Result<TokenStream, Error>

this function outputs all the generated code, it is composed into two different functions: intern_<original_fn_name> parse_<original_fn_name>

intern_ is the simplest function, it is generated to be called within sl-sh to avoid writing boilerplate code to submit a function symbol and the associated code to the runtime.

parse_ has the same function signature as all rust native functions, it takes the environment and a list of args. It evals those arguments at runtime and converts them to rust types they can be consumed by the target rust function.