Function bridge_macros::parse_variadic_args_type

pub(crate) fn parse_variadic_args_type(
    arg_name_itself_is_iter: bool,
    ty: &TypePath,
    fn_name: &str,
    arg_name: &Ident,
    inner: TokenStream,
    collect_type: TokenStream
) -> Result<TokenStream, Error>

if expecting a Vec then the actual expression itself should be iterable (Pair/Vector/Nil), set arg_name_itself_is_iter = true. Otherwise it’s varargs which means the original args vector is passed in and iterated over. the implication of passing in Vec is that if the Exp doesn’t match one of a few underlying sl-sh types passing it to this function is an error.