Crate bridge_macros

TODO PC ISSUE #9 macro crate wish list

The macro should fail if the structure of the docs is not as expected. e.g. Type/Namespace/…/Usage/Example/

type aliasing, need a gensym type macro so I do not conflict with names.

trybuild tests!

worry about inlining, e.g. are the mechanisms in place to do the type conversions constant time, and inlined? or is there a way to make them so?

can yet another crate solve the problem of housing typehandle/passingstyle/param/errorstrings/typedwrapper in the same place?

EVERY macro doesn’t need the check for VarArgs, only signatures that have VarArgs, this can be addressed in the macro.

Port tuple tests: https://github.com/sl-sh-dev/sl-sh/blob/83cd1c93d1eea726ba138a83610155bdbfcab7aa/src/builtins_types.rs#L628

To avoid needing to do lifetimes, if the return value is one of the INPUT values have that marked in the annotation… OTHERWISE data may be copied! OR allow for simple lifetimes?

VMError should not be new_vm

Optional duplicates code. With the ownership model from SlFromRef, being the primary mechanism for crossing the boundary, it’s possible having Some and None blocks multiple times is not necessary.

Support for slices? to avoid Vec allocation? Is it a big deal to only be able to accept Vec? The decision should at least be documented.

SINCE WHEN is it a requirement like that it has to return VMResult or Option

Enums

RustType 🔒

SupportedGenericReturnTypes 🔒

Constants

POSSIBLE_ARG_TYPES 🔒

POSSIBLE_RETURN_TYPES 🔒

SPECIAL_ARG_TYPES 🔒

Functions

are_args_valid 🔒

Optional and VarArgs types are supported to create the idea of items that might be provided or for providing a list of zero or more items that can be passed in. The nature of optional and varargs are context dependent because variable numbers of arguments have to be at the end of the function signature. This method verifies that items marked as Optional are last, and VarArgs is supported but only in the last position, which can be after any number of Optional arguments. This means non Optional/VarArgs types must come before all Optional and VarArgs types.

embed_params_vec 🔒

create a vec literal of the expected Param types so code can check its arguments at runtime for API arity/type correctness.

generate_assertions_code_for_return_type_conversions 🔒

generate_builtin_fn 🔒

write the builtin_ version of the provided function. This function is the function taht makes a direct call to the original rust native function to which the macro was applied. To accomplish this the builtin_ function generates takes some number of ArgType structs (the wrapper enum that enables passing optional and varargs). the body of the function handles unwrapping the ArgType variables and then unwrapping the Expressions those contain into the proper rust native function. The process is done in a for loop but it recursively builds the body of builtin_ by passing around a token stream.

generate_inner_fn_signature_to_orig_fn_call 🔒

create two lists that can be joined by macro syntax to create the inner part of a function signature, e.g. (arg_0: a_type, arg_1: b_type, …) in some existing rust function: fn myfn(arg_0: a_type, arg_1: b_type, …) { … }

generate_intern_fn 🔒

write the intern_ function code. This code is generated to be called within sl-sh to avoid writing boilerplate code to submit a function symbol and the associated code to the runtime. Every builtin function must be inserted into a hashmap where the key is the name of the function and the value is a function expression that stores the name of the rust function to call and its documentation. It looks like the following in all cases:

generate_parse_fn 🔒

write the parse_ version of the provided function. The function it generates takes an environment and a list of Value, evaluates those expressions and then maps the provided list of expressions to a list of ArgType values. To accomplish this information from compile time, arg_types, is manually inserted into this function. This way the evaluated list of args and the expected list of args can be compared and the appropriate vector of arguments can be created and passed to the builtin function. To map a vector of ArgType structs to an actual function call the ExpandVecToArgs trait is used. A sample parse_ function for a function that takes one argument is shown below.

generate_sl_sh_fn 🔒

this function outputs all of the generated code, it is composed into two different functions: intern_<original_fn_name> parse_<original_fn_name>

get_arg_pos 🔒

get_attribute_name_value 🔒

get_attribute_value_with_key 🔒

get_bool_attribute_value_with_key 🔒

get_const_params_len_ident 🔒

get_documentation_for_fn 🔒

Pull out every #doc attribute on the target fn for the proc macro attribute. Ignore any other attributes and only Err if there are no #doc attributes.

get_generic_argument_from_type 🔒

get_generic_argument_from_type_path 🔒

get_intern_fn_name 🔒

return the function names the macro will create. Given a base name, return intern_ Ident to be used as function name

get_param_from_type 🔒

get_parse_fn_name 🔒

return the function names the macro will create. Given a base name, return parse_ Ident to be used as function name

get_parser_for_type_handle 🔒

get_return_type 🔒

returns the option of inner type and the wrapped generic type (None if it’s not generic. If there is no return type None, None is returned. Throws an error if the generic return type is not in the list of predefined constants POSSIBLE_RESULT_TYPES.

get_type_handle 🔒

get_type_or_wrapped_type 🔒

at this point the macro is only operating on types it expects which are any rust types, any rust types wrapped in Option, and any rust types wrapped in Vec. If in the future this is confusing wrapper types can be made, i.e. SlshVarArgs, so normal rust Vec could be used without being turned into a SlshVarArgs

is_valid_generic_type 🔒

is_vec 🔒

None if not Rust type vec

make_orig_fn_call 🔒

generate an (optionally inlined) call to the original fn with the names of all of the variables, arg_names, filled in,, e.g. fn myfn(arg_0, arg_1, arg_2, ...) { ... }. This function also inserts a dynamic check to throw an error if too many arguments were provided based on the signature of the target function.

no_parse_param 🔒

num_required_args 🔒

opt_is_valid_generic_type 🔒

parse_attributes 🔒

parse_direct_type 🔒

for regular Value values (no Optional/VarArgs) ref_exp just needs to be matched based on it’s ExpEnum variant.

parse_fn_arg_type 🔒

recursively wrap the received sl_sh args at the given idx in callback functions that parse the Expressions into the a variable with a predefined name with the same type as the corresponding parameter in the rust native function.

parse_optional_type 🔒

for Option values the ref_exp must first be parsed as an Option, and only in the case that the option is Some will it be necessary to match against every ExpEnum variant.

parse_param 🔒

parse_src_function_arguments 🔒

Create a Vec from the original fn’s signature. Information is needed at compile and run time to translate the list of sl_sh expressions to rust native types. This Arg types stores the information about the rust native type (Value/Option/Var) as well as whether it’s moved, passed by reference, or passed by mutable reference.

parse_type 🔒

create the nested match statements to parse rust types into sl_sh types. the rust types will determine what sl_sh functions will be used for transformation. If this function throws errors it means that the inputs, val/passing style are wrong and aren’t matching to the ArgType(s) properly, or the rust type lookup function is busted.

parse_type_tuple 🔒

parse_variadic_args_type 🔒

if expecting a Vec then the actual expression itself should be iterable (Pair/Vector/Nil), set arg_name_itself_is_iter = true. Otherwise it’s varargs which means the original args vector is passed in and iterated over. the implication of passing in Vec is that if the Exp doesn’t match one of a few underlying sl-sh types passing it to this function is an error.

tokens_for_matching_references 🔒

return a code for how to refer to the inner exp enum referent type in an function call.

Type Aliases

MacroResult 🔒

Attribute Macros

sl_sh_fn

macro that creates a bridge between rust native code and sl-sh code, specify the lisp function name to be interned with the “fn_name” attribute. This macro outputs all of the generated bridge code as well as the original function’s code so it can be used by the generated bridge code.