bridge_macros

Function generate_builtin_fn

pub(crate) fn generate_builtin_fn(
    original_item_fn: &ItemFn,
    original_fn_name_str: &str,
    fn_name: &str,
    params: &[Param],
    fn_name_ident: &Ident,
    takes_env: bool,
    inline: bool,
) -> Result<TokenStream, Error>

write the builtin_ version of the provided function. This function is the function that makes a direct call to the original rust native function to which the macro was applied. To accomplish this the builtin_ function generates takes some number of ArgType structs (the wrapper enum that enables passing optional and varargs). the body of the function handles unwrapping the ArgType variables and then unwrapping the Expressions those contain into the proper rust native function. The process is done in a for loop but it recursively builds the body of builtin_ by passing around a token stream.

The token stream is initialized with code to call to the original rust native function with pre-generated names for each of the arguments, e.g. my_rust_native_function(arg_0, arg_1);. Each subsequent iteration of the loop takes the previous token stream returned by the loop and uses that as it’s innermost scope. Thus the original function call is at the core of a series of scopes that create all the necessary arguments with the proper types that were specified on initialization. For a function of one argument that means the code would look something like: